Fix: bullets spawning intersecting with battlefield boundaries.
Bullets always spawned in the exact same position as the turrets, which
means, for large bullets, they would spawn in such a way that they
intersect with the battlefield boundaries, which would cause physics
issues.

Large bullets now spawn with a slight position offset from the turret to
ensure they fit within the battlefield on spawn.