Fix: workaround the issue with despawning `ParticleEffectBundle`.
Despawning a bunch of `ParticleEffectBundle` at the same time seem to
cause immense lag to the point where everything gets almost frozen.

This commit works around the issue by assigning the color
`LinearRgba::NONE` to the trail effect properties when the associated
`WorkerBall` gets despawned, effectively hiding the effect instead of
actually despawning it.

This commit also updates the `EffectAsset` for the trail effect. The
head (group 0) is now initialized with the color `LinearRgba::NONE`,
so it's always invisible. Only the cloned particles become visible
through the update color modifier.