Add: ability to restart after a game is over.
This implementation is not ideal. Upon firing of the `RestartEvent`,
almost all entities are deleted and added again. This causes a big lag
and the game would freeze for a second or two.

Also, any particle effect still cannot be despawned. Needed to reuse
worker trail entities after restarting. Also also, to avoid having
trails crossing the screen from wherever it was left after the original
worker ball got despawned to wherever the new worker ball got spawned
in, balls are preemptively moved to the top of the panels on the left
and right side in equal amount upon restart. Then an initial delay of
`TRAIL_LIFETIME` is added to the worker ball spawn timer to allow the
old trail to fade completely.